refactor: split RoundTracker.test.tsx into focused test files
- Create shared-test-utils.ts for common test utilities
- Split into 8 focused test files based on functionality:
  - RoundTracker.counter.test.tsx (round counter controls)
  - RoundTracker.duration.test.tsx (duration tracking)
  - RoundTracker.effects.test.tsx (effects management)
  - RoundTracker.history.test.tsx (round history)
  - RoundTracker.triggers.test.tsx (triggers and reminders)
  - RoundTracker.export.test.tsx (export and summary)
  - RoundTracker.accessibility.test.tsx (accessibility)
  - RoundTracker.performance.test.tsx (performance)
  - RoundTracker.error.test.tsx (error handling)
- Remove original 496-line file to comply with 450-line limit
- All tests continue to pass with improved organization
